refactor(Account): move forms under components directory
